Move ConditionalWrapper to lib
