+++ apps/spruce/src/components/HistoryTable/Cell/index.tsx
@@ -3,8 +3,8 @@ import { palette } from "@leafygreen-ui/palette";
 import Tooltip from "@leafygreen-ui/tooltip";
 import { Skeleton } from "antd";
 import { Link } from "react-router-dom";
+import ConditionalWrapper from "@evg-ui/lib/components/ConditionalWrapper";
 import { TaskStatus } from "@evg-ui/lib/types/task";
-import { ConditionalWrapper } from "components/ConditionalWrapper";
 import { inactiveElementStyle, StyledRouterLink } from "components/styles";
 import { getTaskRoute } from "constants/routes";
 import { size } from "constants/tokens";
