+++ apps/spruce/src/components/TaskStatusBadge/index.tsx
@@ -3,8 +3,8 @@ import styled from "@emotion/styled";
 import Badge, { Variant } from "@leafygreen-ui/badge";
 import { palette } from "@leafygreen-ui/palette";
 import { Link } from "react-router-dom";
+import ConditionalWrapper from "@evg-ui/lib/components/ConditionalWrapper";
 import { TaskStatus } from "@evg-ui/lib/types/task";
-import { ConditionalWrapper } from "components/ConditionalWrapper";
 import { getTaskRoute } from "constants/routes";
 import { taskStatusToCopy } from "constants/task";
 import { TaskTab } from "types/task";
