+++ apps/spruce/src/components/TreeSelect/TreeSelect.tsx
@@ -2,7 +2,7 @@ import { useEffect } from "react";
 import styled from "@emotion/styled";
 import Checkbox from "@leafygreen-ui/checkbox";
 import { palette } from "@leafygreen-ui/palette";
-import { ConditionalWrapper } from "components/ConditionalWrapper";
+import ConditionalWrapper from "@evg-ui/lib/components/ConditionalWrapper";
 import { FilterInputControls } from "components/FilterInputControls";
 import { tableInputContainerCSS } from "components/styles/Table";
 import { size } from "constants/tokens";
