+++ apps/spruce/src/pages/spawn/spawnVolume/spawnVolumeTableActions/UnmountButton.tsx
@@ -1,8 +1,8 @@
 import { useQuery, useMutation } from "@apollo/client";
 import Button, { Size } from "@leafygreen-ui/button";
 import Tooltip from "@leafygreen-ui/tooltip";
+import ConditionalWrapper from "@evg-ui/lib/components/ConditionalWrapper";
 import { useSpawnAnalytics } from "analytics/spawn/useSpawnAnalytics";
-import { ConditionalWrapper } from "components/ConditionalWrapper";
 import Popconfirm from "components/Popconfirm";
 import { useToastContext } from "context/toast";
 import {
